requirements.txt
Django==1.8
PyYAML==3.11
altgraph==0.12
argh==0.26.1
cffi==0.9.2
cryptography==0.8.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.1
django-appconf==1.0.1
django-braces==1.4.0
django-crispy-forms==1.4.0
django-oauth-toolkit==0.8.0
django-oauth2-provider==0.2.6.1
django-rest-auth==0.3.4
djangorestframework==3.1.1
enum34==1.0.4
httplib2==0.9
macholib==1.7
oauthlib==0.7.2
pathtools==0.1.2
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.10
python-openid==2.2.5
requests==2.6.0
requests-oauthlib==0.4.2
shortuuid==0.4.2
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
